refactor(pos-event-receiver): move EventDao into internal.dao (#1549)
EventDao sat outside internal in a public dao package while its
signatures import internal.dto and internal.entity types — the leak
shape ADR-0026 D4 forbids. pos-event-receiver holds no cross-module
grant, so nothing belongs outside internal except the application
class. Move the interface to internal.dao alongside EventDaoImpl and
update imports; no signature or behavior changes.

Closes #1549
