+++ pos-event-receiver/src/test/java/com/positivity/poseventreceiver/internal/service/EmitEventServiceImplTest.java
@@ -8,7 +8,7 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.positivity.poseventreceiver.dao.EventDao;
+import com.positivity.poseventreceiver.internal.dao.EventDao;
 import com.positivity.poseventreceiver.internal.dto.EmitEventRequest;
 import java.time.Instant;
 import org.junit.jupiter.api.DisplayName;
